Move all files under kotlin directory
