+++ app/src/main/kotlin/kt/school/starlord/App.kt
