+++ app/src/main/kotlin/kt/school/starlord/AppActivity.kt
