+++ app/src/main/kotlin/kt/school/starlord/model/network/NetworkRepository.kt
