+++ app/src/main/kotlin/kt/school/starlord/model/room/entity/RoomCategory.kt
