+++ app/src/main/kotlin/kt/school/starlord/ui/categories/CategoryAdapterDelegate.kt
