+++ app/src/test/kotlin/kt/school/starlord/ui/categories/CategoriesFragmentTest.kt
